refactor: reduce duplication in relations manager
